Refactor metrics: move EditToolbar
