+++ app/models/Metrics.js
@@ -1,0 +1,20 @@
+const MetricsModel = [
+    {
+        field: "name",
+        headerName: "Name",
+        flex: 1,
+        editable: true,
+    },
+    {
+        field: "description",
+        headerName: "Description",
+        flex: 3,
+        editable: true,
+    },
+    {
+        field: "unit",
+        headerName: "Unit",
+        flex: 1,
+        editable: true,
+    },
+];
